% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_wkt.R
\name{as_wkt}
\alias{as_wkt}
\title{Convert to WKT}
\usage{
as_wkt(x, fmt = 16, ...)
}
\arguments{
\item{x}{Output from \code{\link{region_geojson}}}

\item{fmt}{Format string which indicates the number of digits to display after the
decimal point when formatting coordinates.}

\item{...}{Further args passed on to \code{\link[jsonlite]{fromJSON}} only in the event of json
passed as a character string.}
}
\description{
Convert to WKT
}
\examples{
\dontrun{
res <- region_geojson(key = "MarineRegions:eez_33176")
as_wkt(res, fmt = 5)

nms <- region_names()
res <- region_geojson(key = nms[[1]]$name)
as_wkt(res, fmt = 5)
}
}

