% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mregions-package.R
\docType{package}
\name{mregions-package}
\alias{mregions}
\alias{mregions-package}
\title{Marine regions data}
\description{
Marine regions data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

