% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis_eez_id.R
\name{obis_eez_id}
\alias{obis_eez_id}
\title{Get OBIS EEZ id}
\usage{
obis_eez_id(x)
}
\arguments{
\item{x}{Output from \code{\link{region_geojson}}}
}
\description{
Get OBIS EEZ id
}
\examples{
\dontrun{
(res <- region_names())
obis_eez_id(x = res[[1]]$title)
}
}

