% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_relations.R
\name{place_relations}
\alias{place_relations}
\title{Related records}
\usage{
place_relations(mrgid, direction = c("upper", "lower", "both"),
  type = c("partof", "partlypartof", "adjacentto", "similarto",
  "administrativepartof", "influencedby", "all"), ...)
}
\arguments{
\item{mrgid}{the MRGID (Marineregions Global Identifier) for the
record of interest}

\item{direction}{in which direction of the geographical hierarchy
should the records be retrieved?}

\item{type}{what kind of relations should the records retrieve
have with the place?}

\item{...}{curl options to be passed on to \code{\link[httr]{GET}}}
}
\description{
Get related records based on their MRGID.
}
\examples{
\dontrun{
  tikehau <- geo_code("tikehau")
  place_relations(tikehau$MRGID)
}
}
\author{
Francois Michonneau <francois.michonneau@gmail.com>
}

