% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_types.R
\name{place_types}
\alias{place_types}
\title{Get Marineregions place types}
\usage{
place_types(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A data.frame with the columns:
\itemize{
 \item type (character) the place type
 \item description (character) description of the place type
}
}
\description{
Get Marineregions place types
}
\examples{
\dontrun{
res <- place_types()
head(res)
res$type
}
}

