% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Get a region shp file}
\usage{
region(name = NULL, key = NULL, format = "geojson")
}
\arguments{
\item{name}{(character) Region name}

\item{key}{(character) Region key}

\item{format}{(character) One of 'geojson' or 'shp'}
}
\description{
Get a region shp file
}
\examples{
\dontrun{
# by key
res <- region(key = "MarineRegions:eez_33176", format = "geojson")

# by name
res <- region(name = "Turkmen Exclusive Economic Zone", format = "geojson")

library("geojsonio")
as.json(res) \%>\% map_leaf

nms <- region_names()
as.json(region(nms[[1]]$name)) \%>\% map_leaf
}
}

