% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_geojson.R
\name{region_geojson}
\alias{region_geojson}
\title{Get a Marineregions geojson file}
\usage{
region_geojson(name = NULL, key = NULL, maxFeatures = 50, ...)
}
\arguments{
\item{name}{(character) Region name}

\item{key}{(character) Region key}

\item{maxFeatures}{(integer) Number of features}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get a Marineregions geojson file
}
\examples{
\dontrun{
# by key
res <- region_geojson(key = "MarineRegions:eez_33176")

# by name
res <- region_geojson(name = "Turkmen Exclusive Economic Zone")

library("geojsonio")
as.json(res) \%>\% map_leaf

nms <- region_names()
as.json(region_geojson(nms[[1]]$name)) \%>\% map_leaf
}
}

