% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_names.R
\name{region_names}
\alias{region_names}
\title{Get region names}
\usage{
region_names(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
a data.frame, or tibble, of class tbl_df (basically, a compact data.frame),
with slots:
\itemize{
 \item name (character) - name of the region, which is a combination of the
 name_first and name_second, e.g., Morocco:elevation_10m
 \item title (character) - title for the region
 \item name_first (character) - first part of the name, e.g., Morocco
 \item name_second (character) - second part of the name, e.g., elevation_10m
}
}
\description{
Get region names
}
\examples{
\dontrun{
(res <- region_names())
res$name
sort(unique(res$name_first))
sort(unique(res$name_second))
res$title
}
}

