% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_names_search.R
\name{region_names_search}
\alias{region_names_search}
\title{Search for region names}
\usage{
region_names_search(x, q = NULL, ...)
}
\arguments{
\item{x, q}{Either a \code{tbl_df}, returned from \code{\link{region_names}}, or
a query as a character string. If a \code{tbl_df}, you must pass a query string to
\code{q}. If a query string (character) is passed to \code{x}, leave \code{q}
as \code{NULL}}

\item{...}{Parameters passed on to \code{\link{agrep}}}
}
\value{
a data.frame, or tibble, of class tbl_df
}
\description{
Search for region names
}
\examples{
\dontrun{
(res <- region_names())

region_names_search(res, q = "EEZ")
region_names_search("IHO")
region_names_search(res, q = "IHO")
region_names_search("Heritage")
region_names_search(res, q = "Heritage")
region_names_search("ecoregions")
region_names_search(res, q = "ecoregions")
}
}

