% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis_eez_id.R
\name{obis_eez_id}
\alias{obis_eez_id}
\title{Get OBIS EEZ id}
\usage{
obis_eez_id(x)
}
\arguments{
\item{x}{(character) An Exclusive Economic Zone name}
}
\value{
An integer EEZ ID if a match found in list of EEZ's, or \code{NULL} if
no match found.
}
\description{
Get OBIS EEZ id
}
\examples{
\dontrun{
(res <- region_names())
obis_eez_id(res$title[100])
}
}

