% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records_by_type.R
\name{records_by_type}
\alias{records_by_type}
\title{Get Marineregions records by place type}
\usage{
records_by_type(type, offset = 0, ...)
}
\arguments{
\item{type}{(character) One place type name. See \code{\link{place_types}} for
place type names}

\item{offset}{(numeric) Offset to start at. Each request can return up to
100 results. e.g., an offset of 200 will give records 200 to 299.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
If no results, an empty list. If results found, a data.frame with the columns:
\itemize{
 \item MRGID (integer)
 \item gazetteerSource (character)
 \item placeType (character)
 \item latitude (numeric)
 \item longitude (numeric)
 \item minLatitude (numeric)
 \item minLongitude (numeric)
 \item maxLatitude (numeric)
 \item maxLongitude (numeric)
 \item precision (numeric)
 \item preferredGazetteerName (character)
 \item preferredGazetteerNameLang (character)
 \item status (character)
 \item accepted (integer)
}
}
\description{
Get Marineregions records by place type
}
\details{
Internally we use the \code{getGazetteerRecordsByType.json} API method,
which searches for Marineregions records by user supplied place type
}
\examples{
\dontrun{
res <- records_by_type(type="EEZ")
head(res)

types <- place_types()
records_by_type(types$type[1])
records_by_type(types$type[10])
records_by_type(grep("MEOW", types$type, value = TRUE))
records_by_type(grep("MEOW", types$type, value = TRUE), offset = 100)
}
}

