% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Get a region shp file}
\usage{
region(x, format = "geojson")
}
\arguments{
\item{x}{Region name}

\item{format}{one of geojson, shp}
}
\description{
Get a region shp file
}
\examples{
\dontrun{
res <- region(x = "MarineRegions:eez_33176", format = "geojson")

library("geojsonio")
as.json(res) \%>\% map_leaf

nms <- region_names()
as.json(region(nms[[1]]$name)) \%>\% map_leaf
}
}

