% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_names.R
\name{region_names}
\alias{region_names}
\title{Get region names}
\usage{
region_names(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
a data.frame, or tibble, of class tbl_df
}
\description{
Get region names
}
\examples{
\dontrun{
(res <- region_names())
res$name
sort(unique(res$name_first))
sort(unique(res$name_second))
res$title
}
}

