% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_names_search.R
\name{region_names_search}
\alias{region_names_search}
\title{Search for region names}
\usage{
region_names_search(x, q = NULL, ...)
}
\arguments{
\item{x, q}{Either a \code{tbl_df}, returned from \code{\link{region_names}}, or
a query as a character string. If a \code{tbl_df}, you must pass a query string to
\code{q}. If a query string (character) is passed to \code{x}, leave \code{q}
as \code{NULL}}

\item{...}{Parameters passed on to \code{\link{agrep}}}
}
\value{
\code{NULL} if no matches found, or a data.frame, or tibble, of class
tbl_df (basically, a compact data.frame), with slots:
\itemize{
 \item name (character) - name of the region, which is a combination of the
 name_first and name_second, e.g., Morocco:elevation_10m
 \item title (character) - title for the region
 \item name_first (character) - first part of the name, e.g., Morocco
 \item name_second (character) - second part of the name, e.g., elevation_10m
}
}
\description{
Search for region names
}
\examples{
\dontrun{
# Get region names with region_names() function
(res <- region_names())

# to save time, pass in the result from region_names()
region_names_search(res, q = "EEZ")

# if you don't pass in the result from region_names(), we have to
# call region_names() internally, adding some time
region_names_search("IHO")

# more examples, with and without passing in region_names() output
region_names_search(res, q = "IHO")
region_names_search("Heritage")
region_names_search(res, q = "Heritage")
region_names_search("ecoregions")
region_names_search(res, q = "ecoregions")
}
}

