% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region_shp}
\alias{region_geojson}
\alias{region_shp}
\title{Get a region shp file}
\usage{
region_shp(name = NULL, key = NULL, maxFeatures = 50,
  path = "~/.mregions", overwrite = TRUE, read = TRUE)

region_geojson(name = NULL, key = NULL, maxFeatures = 50)
}
\arguments{
\item{name}{(character) Region name}

\item{key}{(character) Region key}

\item{maxFeatures}{(integer) Number of features}

\item{path}{(character) path to write shp files to, only used in \code{region_shp}}

\item{overwrite}{(logical) Overwrite file if already exists. Default: \code{FALSE}}

\item{read}{(logical) To read in as spatial object. If \code{FALSE} a path
given back. if \code{TRUE}, you need the \code{rgdal} package installed.
Default: \code{FALSE}}
}
\description{
Get a region shp file
}
\examples{
\dontrun{
# by key
res <- region_geojson(key = "MarineRegions:eez_33176")

# by name
res <- region_geojson(name = "Turkmen Exclusive Economic Zone")

library("geojsonio")
as.json(res) \%>\% map_leaf

nms <- region_names()
as.json(region_geojson(nms[[1]]$name)) \%>\% map_leaf

# shp files
## just get path
region_shp(key = "MarineRegions:eez_33176", read = FALSE)
## read shp file into spatial object
res <- region_shp(key = "MarineRegions:eez_33176", read = TRUE)
library('leaflet')
leaflet() \%>\% addPolygons(data = res)
}
}

