% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geo_code.R
\name{rev_geo_code}
\alias{rev_geo_code}
\title{Reverse Geocode with Marineregions}
\usage{
rev_geo_code(lat, lon, lat_radius = 1, lon_radius = 1, ...)
}
\arguments{
\item{lat}{Latitude for the coordinates (decimal format)}

\item{lon}{Longitude for the coordinates (decimal format)}

\item{lat_radius}{Extends search to include the range from
\code{lat}-\code{lat_radius} to \code{lat}+\code{lat_radius}}

\item{lon_radius}{Extends search to include the range from
\code{lon}-\code{lon_radius} to \code{lon}+\code{lon_radius}}

\item{...}{curl options to be passed on to \code{\link[httr]{GET}}}
}
\description{
Retrieve the names of geographic objects from coordinates (and
optionally a radius around them).
}
\examples{
\dontrun{
    rev_geo_code(-21.5, 55.5, lat_radius=.5, lon_radius=.5)
}
}

